Switch gelf include to libelf include
This commit avoid the use of the gelf header, in favor of the libelf one.
The library is still compatible with 32 or 64 bits architectures (the 32
bits mode is enabled by default).

 * src/mppa_dl.c, include/mppa_dl.h (mppa_dl_load): no need of the image
size anymore.
 * src/mppa_dl_error.c, include.mppa_dl_error.h: use a macro for verbose
modes and clean up error messages.
 * include/mppa_dl_types.h: new handle type, new ElfK1_* types, and add
a relocation type.
 * src/mppa_dl_load.c, include/mppa_dl_load.h: no need of the gelf*
functions.

Plus a lot of code cleaning and some comments.