Sync the mppa_dl_debug struct for ClusterOS
Ref T10284

Summary:
For multiclusters SMP on ClusterOS, mppa_dl_struct
resides in SMEM and must be synchronized for all SMP clusters in
order to provide the same so library list. In addition, a valid
field was added to the mppa_dl_struct struct so that gdb can know
when the so list is modified and it cannot be accessed.

Test Plan: Tested by the opencl testsuite

Reviewers: O24 libmppadl!

Differential Revision: https://phab.kalray.eu/D2384