Add --no-dynamic-linker linker flags
Summary:
Newer binutils are checking that dynamic executable have the ELF program headers
loaded or raise an error. In our case, we don't want them and can't easily add
them. They are used by the dynamic loader that is supposed to load the resulting
program, which is not our case. We are using dynamic linking to get .dynamic
sections but the program is self contained and does not need an external dynamic
linker.

ref T13792

Test Plan: CI

Reviewers: O24 libmppadl!

Maniphest Tasks: T13792

Differential Revision: https://phab.kalray.eu/D8695