Add support for self-contained tracing
Summary:
When a dynamic library is loaded, this commit calls a function
which sends a tracepoint containing the trace metadata of the loaded
library. Using this info, the trace parsers will be then able to decode
the trace data without requiring the library file.

Ref T16932

Test Plan: The integratin tests should continue to pass

Reviewers: O24 libmppadl!

Differential Revision: https://phab.kalray.eu/D14383